+++ codec/src/test/java/io/netty/handler/codec/ByteToMessageDecoderTest.java
@@ -15,11 +15,7 @@
  */
 package io.netty.handler.codec;
 
-import io.netty.buffer.ByteBuf;
-import io.netty.buffer.CompositeByteBuf;
-import io.netty.buffer.Unpooled;
-import io.netty.buffer.UnpooledByteBufAllocator;
-import io.netty.buffer.UnpooledHeapByteBuf;
+import io.netty.buffer.*;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
 import io.netty.channel.ChannelOutboundHandlerAdapter;
@@ -50,7 +46,7 @@ public class ByteToMessageDecoderTest {
             }
         });
 
-        ByteBuf buf = Unpooled.wrappedBuffer(new byte[] {'a', 'b', 'c'});
+        ByteBuf buf = Unpooled.wrappedBuffer(new byte[]{'a', 'b', 'c'});
         channel.writeInbound(buf.copy());
         ByteBuf b = channel.readInbound();
         assertEquals(b, buf.skipBytes(1));
@@ -60,7 +56,7 @@ public class ByteToMessageDecoderTest {
 
     @Test
     public void testRemoveItselfWriteBuffer() {
-        final ByteBuf buf = Unpooled.buffer().writeBytes(new byte[] {'a', 'b', 'c'});
+        final ByteBuf buf = Unpooled.buffer().writeBytes(new byte[]{'a', 'b', 'c'});
         EmbeddedChannel channel = new EmbeddedChannel(new ByteToMessageDecoder() {
             private boolean removed;
 
@@ -77,7 +73,7 @@ public class ByteToMessageDecoderTest {
         });
 
         channel.writeInbound(buf.copy());
-        ByteBuf expected = Unpooled.wrappedBuffer(new byte[] {'b', 'c'});
+        ByteBuf expected = Unpooled.wrappedBuffer(new byte[]{'b', 'c'});
         ByteBuf b = channel.readInbound();
         assertEquals(expected, b);
         expected.release();
@@ -91,7 +87,7 @@ public class ByteToMessageDecoderTest {
      */
     @Test
     public void testInternalBufferClearReadAll() {
-        final ByteBuf buf = Unpooled.buffer().writeBytes(new byte[] {'a'});
+        final ByteBuf buf = Unpooled.buffer().writeBytes(new byte[]{'a'});
         EmbeddedChannel channel = newInternalBufferTestChannel();
         assertFalse(channel.writeInbound(buf));
         assertFalse(channel.finish());
@@ -103,11 +99,11 @@ public class ByteToMessageDecoderTest {
      */
     @Test
     public void testInternalBufferClearReadPartly() {
-        final ByteBuf buf = Unpooled.buffer().writeBytes(new byte[] {'a', 'b'});
+        final ByteBuf buf = Unpooled.buffer().writeBytes(new byte[]{'a', 'b'});
         EmbeddedChannel channel = newInternalBufferTestChannel();
         assertTrue(channel.writeInbound(buf));
         assertTrue(channel.finish());
-        ByteBuf expected = Unpooled.wrappedBuffer(new byte[] {'b'});
+        ByteBuf expected = Unpooled.wrappedBuffer(new byte[]{'b'});
         ByteBuf b = channel.readInbound();
         assertEquals(expected, b);
         assertNull(channel.readInbound());
@@ -162,7 +158,7 @@ public class ByteToMessageDecoderTest {
     @Test
     public void testFireChannelReadCompleteOnInactive() throws InterruptedException {
         final BlockingQueue<Integer> queue = new LinkedBlockingDeque<Integer>();
-        final ByteBuf buf = Unpooled.buffer().writeBytes(new byte[] {'a', 'b'});
+        final ByteBuf buf = Unpooled.buffer().writeBytes(new byte[]{'a', 'b'});
         EmbeddedChannel channel = new EmbeddedChannel(new ByteToMessageDecoder() {
             @Override
             protected void decode(ChannelHandlerContext ctx, ByteBuf in, List<Object> out) throws Exception {
@@ -225,7 +221,7 @@ public class ByteToMessageDecoderTest {
             }
         });
 
-        ByteBuf buf = Unpooled.wrappedBuffer(new byte[] { 'a', 'b', 'c' });
+        ByteBuf buf = Unpooled.wrappedBuffer(new byte[]{'a', 'b', 'c'});
         assertTrue(channel.writeInbound(buf.copy()));
         ByteBuf b = channel.readInbound();
         assertEquals(b, buf.skipBytes(1));
@@ -304,7 +300,7 @@ public class ByteToMessageDecoderTest {
             }
         });
         assertFalse(channel.writeInbound(Unpooled.buffer(8).writeByte(1).asReadOnly()));
-        assertFalse(channel.writeInbound(Unpooled.wrappedBuffer(new byte[] { (byte) 2 })));
+        assertFalse(channel.writeInbound(Unpooled.wrappedBuffer(new byte[]{(byte) 2})));
         assertFalse(channel.finish());
     }
 
@@ -367,27 +363,27 @@ public class ByteToMessageDecoderTest {
         assertEquals(0, interceptor.readsTriggered);
 
         // 0 complete frames, 1 partial frame: SHOULD trigger a read
-        channel.writeInbound(wrappedBuffer(new byte[] { 0, 1 }));
+        channel.writeInbound(wrappedBuffer(new byte[]{0, 1}));
         assertEquals(1, interceptor.readsTriggered);
 
         // 2 complete frames, 0 partial frames: should NOT trigger a read
-        channel.writeInbound(wrappedBuffer(new byte[] { 2 }), wrappedBuffer(new byte[] { 3, 4, 5 }));
+        channel.writeInbound(wrappedBuffer(new byte[]{2}), wrappedBuffer(new byte[]{3, 4, 5}));
         assertEquals(1, interceptor.readsTriggered);
 
         // 1 complete frame, 1 partial frame: should NOT trigger a read
-        channel.writeInbound(wrappedBuffer(new byte[] { 6, 7, 8 }), wrappedBuffer(new byte[] { 9 }));
+        channel.writeInbound(wrappedBuffer(new byte[]{6, 7, 8}), wrappedBuffer(new byte[]{9}));
         assertEquals(1, interceptor.readsTriggered);
 
         // 1 complete frame, 1 partial frame: should NOT trigger a read
-        channel.writeInbound(wrappedBuffer(new byte[] { 10, 11 }), wrappedBuffer(new byte[] { 12 }));
+        channel.writeInbound(wrappedBuffer(new byte[]{10, 11}), wrappedBuffer(new byte[]{12}));
         assertEquals(1, interceptor.readsTriggered);
 
         // 0 complete frames, 1 partial frame: SHOULD trigger a read
-        channel.writeInbound(wrappedBuffer(new byte[] { 13 }));
+        channel.writeInbound(wrappedBuffer(new byte[]{13}));
         assertEquals(2, interceptor.readsTriggered);
 
         // 1 complete frame, 0 partial frames: should NOT trigger a read
-        channel.writeInbound(wrappedBuffer(new byte[] { 14 }));
+        channel.writeInbound(wrappedBuffer(new byte[]{14}));
         assertEquals(2, interceptor.readsTriggered);
 
         for (int i = 0; i < 5; i++) {
@@ -416,10 +412,10 @@ public class ByteToMessageDecoderTest {
         };
         EmbeddedChannel channel = new EmbeddedChannel(decoder);
         assertTrue(channel.writeInbound(Unpooled.wrappedBuffer(new byte[]{1, 2, 3, 4, 5})));
-        assertEquals((byte) 1,  channel.readInbound());
-        assertEquals((byte) 2,  channel.readInbound());
-        assertEquals((byte) 3,  channel.readInbound());
-        assertEquals((byte) 4,  channel.readInbound());
+        assertEquals((byte) 1, (byte) channel.readInbound());
+        assertEquals((byte) 2, (byte) channel.readInbound());
+        assertEquals((byte) 3, (byte) channel.readInbound());
+        assertEquals((byte) 4, (byte) channel.readInbound());
         ByteBuf buffer5 = channel.readInbound();
         assertEquals((byte) 5, buffer5.readByte());
         assertFalse(buffer5.isReadable());
